+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - `page.document` must return `undefined` if the page is not a HTML page [#163].
-- Updated `liquidjs`.
+- Updated `liquidjs`, `deno_graph` and `std`.
 
 ## [1.4.3] - 2022-01-12
 ### Fixed